requirements.txt
alembic==1.4.3
astroid==2.4.2
Babel==2.8.0
blinker==1.4
Brotli==1.0.9
click==7.1.2
dash==1.16.2
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
dnspython==2.0.0
email-validator==1.1.1
flake8==3.8.3
Flask==1.1.2
Flask-Admin==1.5.6
Flask-BabelEx==0.9.4
Flask-Compress==1.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Security-Too==3.4.4
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
idna==2.10
importlib-metadata==2.0.0
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.2
plotly==4.10.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
retrying==1.3.3
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.19
toml==0.10.1
typed-ast==1.4.1
uuid==1.30
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.2.0
